Non-matching decomp of func_80014180_14D80

+++ include/functions.us.h
@@ -65,7 +65,7 @@ s32 func_80011B14_12714();
 s32 func_80011B60_12760();
 void func_80013324_13F24(void);
 void func_80013684_14284(void);
-void func_80014180_14D80(s32 arg0);
+void func_80014180_14D80(s8 arg0);
 void func_80014A3C_1563C(s32 arg0, s16 arg1, f32 arg2, s16 arg3, f32 arg4);
 void func_800153D8_15FD8(s16);
 void func_800157D4_163D4(s16);
